chore(cli): move egon to own module
